Added save image to blog
