+++ src/api/blog/dto/create-blog.dto.ts
@@ -1,4 +1,5 @@
 export class CreateBlogDto {
   readonly title: string;
   readonly text: string;
+  readonly image: Express.Multer.File;
 }
