+++ src/app.controller.ts
@@ -1,12 +1,10 @@
-import { Controller, Get } from '@nestjs/common';
-import { AppService } from './app.service';
+import { Controller, Get, Param, Res } from '@nestjs/common';
 
 @Controller()
 export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  getHello(): string {
-    return this.appService.getHello();
+  @Get('/storage/:folder/:image')
+  seeUploadedFile(@Param('folder') folder, @Param('image') image, @Res() res) {
+    const path = './storage/' + folder;
+    return res.sendFile(image, { root: path });
   }
 }
