+++ src/app.module.ts
@@ -2,10 +2,12 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BlogModule } from './api/blog/blog.module';
 import { Connection } from 'typeorm';
-import { Blog } from "./database/entities/blog";
+import { Blog } from './database/entities/blog';
 import { UserModule } from './api/user/user.module';
 import { AuthModule } from './api/auth/auth.module';
-import { User } from "./database/entities/user";
+import { User } from './database/entities/user';
+import { AppController } from './app.controller';
+
 
 @Module({
   imports: [
@@ -23,6 +25,7 @@ import { User } from "./database/entities/user";
     UserModule,
     AuthModule,
   ],
+  controllers: [AppController],
 })
 export class AppModule {
   constructor(private connection: Connection) {}
