+++ src/database/entities/blog.ts
@@ -13,7 +13,7 @@ export class Blog {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column()
+  @Column({ nullable: true })
   image: string;
 
   @Column({ length: 190 })
